restructured some files
