+++ Source/Xamarin/Prism.Forms/Behaviors/NavigationPageDisposeBehavior.cs
@@ -1,4 +1,4 @@
-﻿using Prism.Common;
+﻿using Prism.Navigation;
 using Xamarin.Forms;
 
 namespace Prism.Behaviors
