+++ Source/Xamarin/Prism.Forms/Navigation/PageNavigationRegistry.cs
@@ -1,9 +1,7 @@
-﻿using Prism.Navigation;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Reflection;
 
-namespace Prism.Common
+namespace Prism.Navigation
 {
     public static class PageNavigationRegistry
     {
