+++ Source/Xamarin/Prism.Forms/Prism.Forms.csproj
@@ -57,15 +57,15 @@
     <Compile Include="Behaviors\NavigationPageDisposeBehavior.cs" />
     <Compile Include="Behaviors\TabbedPageActiveAwareBehavior.cs" />
     <Compile Include="Behaviors\MultiPageActiveAwareBehavior{T}.cs" />
-    <Compile Include="Common\PageUtilities.cs" />
+    <Compile Include="Navigation\PageUtilities.cs" />
     <Compile Include="IPlatformInitializer.cs" />
     <Compile Include="Navigation\INavigationPageOptions.cs" />
     <Compile Include="PrismApplicationBase.cs" />
     <Compile Include="Common\ApplicationProvider.cs" />
     <Compile Include="Common\IApplicationProvider.cs" />
     <Compile Include="Navigation\IMasterDetailPageOptions.cs" />
-    <Compile Include="Common\PageNavigationInfo.cs" />
-    <Compile Include="Common\PageNavigationRegistry.cs" />
+    <Compile Include="Navigation\PageNavigationInfo.cs" />
+    <Compile Include="Navigation\PageNavigationRegistry.cs" />
     <Compile Include="Common\UriParsingHelper.cs" />
     <Compile Include="FormsApplication.cs" />
     <Compile Include="Modularity\IModule.cs" />
